Keep all Snyk specific logic in one file
